Do not remove statically defined peers
If a peer for a client is defined statically, either by the config
or cli, then it makes little sense for it to blacklist connections
to that peer.  It contradicts the intention of the config.
The remote peer may still block the client peer to prevent DoS
style issues.  On the other hand it still makes sense to back off
and be polite if there is a continuing problem.

Signed-off-by: Kevin O'Donnell <kevin@blockchaintp.com>